+++ passes/AT012/testdata/src/a/comment_ignore_test.go
@@ -1,0 +1,11 @@
+package a
+
+import (
+	"testing"
+)
+
+//lintignore:AT012
+func TestAccExampleThingCommentIgnore1_Test(t *testing.T) {}
+
+//lintignore:AT012
+func TestAccExampleThingCommentIgnore2_Test(t *testing.T) {}