+++ passes/AT012/testdata/src/a/failing_test.go
@@ -1,0 +1,10 @@
+// want "file contains multiple acceptance test name prefixes"
+package a
+
+import (
+	"testing"
+)
+
+func TestAccExampleThingFailing1_Test(t *testing.T) {}
+
+func TestAccExampleThingFailing2_Test(t *testing.T) {}