+++ passes/AT012/testdata/src/a/passing_test.go
@@ -1,0 +1,9 @@
+package a
+
+import (
+	"testing"
+)
+
+func TestAccExampleThing_Passing1(t *testing.T) {}
+
+func TestAccExampleThing_Passing2(t *testing.T) {}